refactor(app): extract editing/workflow into app_editor.cpp
Move property + AI-script editing, undo/redo, task-tree ops, QSC load/compile/decompile, attach promotion, and game launch into app_editor.cpp. The shared game-process monitor (GameMonitorParam/GameMonitorProc) and HOTKEY_ID_TOGGLE_GAME move to app_internal.h (LaunchGame uses them); the mutable window-subclass globals stay in core. app.cpp core 5791 -> 1376 lines (now split across app_input/view/lookup/editor + app_internal.h). Pure code motion: builds clean, runtime-verified (level 2, 558 objects, no errors).
